move display name service
